Switch to inputs that are boxed vectors of 'Array r' from hTensor
The idea is that Doubles will be represented as zero-rank arrays,
but the bulk of computations will be done on 1 and 2 rank arrays.
'Delta' datatype doesn't need to be extended and 'eval' is unchanged.

This compiles (after hTensor is fixed a little), but the array operations
are wrong or undefined, because various array operations and their
gradients need to be defined first. This may still not work,
because I may be missing something fundamental. Let's see!