Finish the failed engine and MNIST nn definition overhaul to Arrays
OK, I've gone as far as I could --- it compiles and if wired to test harness, would probably produce correct function values, after some debugging. However, some of the gradient definitions are bogus, e.g., for the operation that sums elements of a vector and the operation that produces a vector of n copies of element x. Both of these operations are required for the softMax activation function in MNIST neural network. Unfortunately, I think Delta has to be extended to handle those and probably others.

The hTensor.Array approach in this failed prototype indeed solves the typing problem, even though at the cost of a very error-prone API, because array of all ranks and sizes have the same type. I think this may be a good setup to handle the extension to Arrays, if we choose to make it, sidestepping for a while the issue of expressing sums and products of well-typed vectors, matrices and tensors.